+++ scripts/enums.ts
